% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DEXSeqResultsAlt}
\alias{DEXSeqResultsAlt}
\title{Function to generate DEXSeqResults object}
\usage{
DEXSeqResultsAlt(object)
}
\arguments{
\item{object}{A \code{DEXSeqDataSet} object.}
}
\value{
A \code{DEXSeqResults} object.
}
\description{
Function to generate a DEXSeqDataSet object based on data from
the pasilla package
}
\examples{
data( "pasillaDEXSeqDataSet", package="pasilla" )
dxd <- estimateSizeFactors( dxd )
dxd <- estimateDispersionsAlt( dxd, verbose=FALSE )
dxd <- testForDEUAlt( dxd, reducedModel=~sample+exon, fullModel=~sample+exon+condition:exon, verbose=FALSE )
dxd <- DEXSeqResultsAlt( dxd )

}

