% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimateDispersionsAlt}
\alias{estimateDispersionsAlt}
\title{Function to estimate dispersions for each exon}
\usage{
estimateDispersionsAlt(dxd, BPPARAM = SerialParam(), verbose = TRUE)
}
\arguments{
\item{dxd}{A \code{DEXSeqDataSet} object.}

\item{BPPARAM}{A \code{BiocParallel} instance.}

\item{verbose}{If the function should be verbose or not.}
}
\value{
A \code{DEXSeqDataSet} with dispersion estimations
included. Note that no sharing across exons is included, thus it
is recommended only for large datasets.
}
\description{
Given a DEXSeqDataSet object, it estimates dispersions
for each exon. The functions for shrinkage of dispersions
from DESeq2 are not implemented. Thus, it is recomended only
for comparisons with large numbers of biological replicates.
}
\examples{
data( "pasillaDEXSeqDataSet", package="pasilla" )
dxd <- estimateSizeFactors( dxd )
dxd <- estimateDispersionsAlt( dxd )

}

