% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generateExampleDxd}
\alias{generateExampleDxd}
\title{Function to generate example DEXSeqDataSet}
\usage{
generateExampleDxd()
}
\description{
Function to generate a DEXSeqDataSet object based on data from
the pasilla package
}

