% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{testForDEUAlt}
\alias{testForDEUAlt}
\title{Function to test for differences in exon usage}
\usage{
testForDEUAlt(dxd, reducedModel, fullModel, BPPARAM = SerialParam(),
  verbose = TRUE)
}
\arguments{
\item{dxd}{A \code{DEXSeqDataSet} object.}

\item{reducedModel}{Formula with the reduced model}

\item{fullModel}{Formula with the full model}

\item{BPPARAM}{A \code{BiocParallel} instance.}

\item{verbose}{If the function should be verbose or not.}
}
\value{
A \code{DEXSeqDataSet} with p-values and q-values.
}
\description{
This function implements testing for differences in exon
usage across biological conditions. It is an alternative
implementation of the original statistical method implemented
in the DEXSeq package. The only difference is that it uses
the stamod glmnb.fit instead of the DESeq2 glm fitter.
}
\examples{
data( "pasillaDEXSeqDataSet", package="pasilla" )
dxd <- estimateSizeFactors( dxd )
dxd <- estimateDispersionsAlt( dxd )

}

